[REFACTOR] 엔티티들 위치 변경
