+++ src/main/java/com/sports/server/command/league/domain/LeagueTopScorer.java
@@ -1,7 +1,6 @@
-package com.sports.server.command.newEntity.league.domain;
+package com.sports.server.command.league.domain;
 
-import com.sports.server.command.league.domain.League;
-import com.sports.server.command.newEntity.player.domain.Player;
+import com.sports.server.command.player.domain.Player;
 import com.sports.server.common.domain.BaseEntity;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
